+++ src/main/kotlin/io/github/teamcheeze/plum/api/core/entity/Movable.kt
@@ -2,6 +2,7 @@ package io.github.teamcheeze.plum.api.core.entity
 
 import org.bukkit.Location
 import org.bukkit.util.Vector
+import java.util.concurrent.CompletableFuture
 
 /**
  * A movable entity
@@ -11,12 +12,12 @@ interface Movable<T>: BaseEntity {
     /**
      * Move a given length
      */
-    fun move(delta: Vector, tick: (Location) -> Unit = {}): T
+    fun move(delta: Vector, speed: Double = 1.0, tick: (Location) -> Unit = {}): CompletableFuture<Location>
 
     /**
      * Move to a given point
      */
-    fun moveTo(target: Location, tick: (Location) -> Unit = {}): T
+    fun moveTo(target: Location, speed: Double = 1.0, tick: (Location) -> Unit = {}): CompletableFuture<Location>
 
     /**
      * Teleport to a given point